Split schema out into versioned submodules to make it more clear how it will be extended in the future
